+++ coders/jpeg.c
@@ -2055,7 +2055,7 @@ static void WriteProfile(j_compress_ptr jpeg_info,Image *image,
         id=JPEG_APP0+StringToInteger(name+3);
         for (i=0; i < (ssize_t) length; i+=65533L)
            jpeg_write_marker(jpeg_info,id,GetStringInfoDatum(profile)+i,
-             MagickMin(length-i,65533));
+             (unsigned int) MagickMin(length-i,65533));
       }
     if (LocaleCompare(name,"EXIF") == 0)
       {
